Add scope for CloudContactBookContext
