+++ Cloud Customer Contact Book/Cloud Customer Contact Book.csproj
@@ -12,4 +12,8 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\CloudDatabase\CloudDatabase.csproj" />
+  </ItemGroup>
+
 </Project>
