+++ Cloud Customer Contact Book/Models/GroupModel.cs
@@ -11,7 +11,7 @@ namespace Cloud_Customer_Contact_Book.Models
         /// </summary>
         /// <example>Family</example>
         [Required]
-        public object Name { get; internal set; }
+        public string Name { get; internal set; }
         /// <summary>
         /// Group identifier.
         /// </summary>
