+++ Cloud Customer Contact Book/WebApi.cs
@@ -2,10 +2,12 @@ using System.Reflection;
 using CloudDatabase;
 using Microsoft.EntityFrameworkCore;
 using WebApi.Services;
+
 var builder = WebApplication.CreateBuilder(args);
 // Add services to the container.
 builder.Services.AddDbContext<CloudContactBookContext>(x => x.UseInMemoryDatabase("ContactBook"));
 builder.Services.AddScoped<ContactService>();
+builder.Services.AddScoped<GroupService>();
 builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
@@ -30,6 +32,9 @@ app.UseHttpsRedirection();
 app.UseAuthorization();
 
 app.MapControllers();
-var context = app.Services.GetRequiredService<CloudContactBookContext>();
-context.Database.EnsureCreated();
+using (var scope = app.Services.CreateScope())
+{
+    var context = scope.ServiceProvider.GetRequiredService<CloudContactBookContext>();
+    context.Database.EnsureCreated();
+}
 app.Run();
