+++ CloudDatabase/Tables/Contact.cs
@@ -10,6 +10,8 @@ namespace CloudDatabase.Tables
     public class Contact
     {
         [Key]
+        public long Id { get; set; }
+
         public string FirstName { get; set; }
 
         public string LastName { get; set; }
