+++ CloudDatabase/Tables/ContactInGroup.cs
@@ -2,9 +2,6 @@
 
 public class ContactInGroup
 {
-    // [Key(0)]
     public long ContactId { get; set; }
-
-    // [Key(1)]
     public long GroupId { get; set; }
 }
