+++ CloudDatabase/Tables/Group.cs
@@ -8,4 +8,7 @@ public class Group
     public long Id { get; set; }
 
     public string Name { get; set; }
+    
+    public int MemberCount { get; set; }
+
 }
