--- Database/Tables/ContactInGroup.cs
@@ -1,10 +1,0 @@
-﻿namespace Database.Tables;
-
-public class ContactInGroup
-{
-    // [Key(0)]
-    public long ContactId { get; set; }
-
-    // [Key(1)]
-    public long GroupId { get; set; }
-}
