--- Database/Tables/Group.cs
@@ -1,11 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Database.Tables;
-
-public class Group
-{
-    [Key]
-    public long Id { get; set; }
-
-    public string Name { get; set; }
-}
